+++ src/cmd/hist.rs
@@ -19,11 +19,23 @@ const MEDIUM_BAR_CHARS: [&str; 1] = ["■"];
 const LARGE_BAR_CHARS: [&str; 8] = ["▏", "▎", "▍", "▌", "▋", "▊", "▉", "█"];
 
 static USAGE: &str = "
-Print a horizontal histogram for the given CSV file with each line
-representing a bar in the resulting graph.
+Print horizontal histograms for the given CSV file with each row representing
+a bar in the graph.
 
-This command is very useful when used in conjunction with the `frequency` or `bins`
-command.
+This command is often used with the `frequency` or `bins` commands:
+
+    $ xan freq -s username tweets.csv | xan hist
+
+    $ xan bins -s retweet_count,like_count tweets.csv | xan hist
+
+Else, this command expects a CSV file with the following columns:
+
+    - \"field\" (optional): the name of a histogram to print
+    - \"value\": a single bar's label
+    - \"count\": a single bar's represented numerical value
+
+You can always customize those column names through the -f/--field, -l/--label
+and -v/--value flags respectively.
 
 Usage:
     xan hist [options] [<input>]
@@ -117,12 +129,20 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
     let mut rdr = conf.reader()?;
     let headers = rdr.byte_headers()?.clone();
 
+    let err_msg = |err: String| {
+        format!("{}\nxan hist expects a field?,value,count CSV input (typically produced by `xan freq` or `xan bins`)!\nSee xan hist --help for more info.", err)
+    };
+
     let label_pos = args
         .flag_label
-        .single_selection(&headers, !conf.no_headers)?;
+        .single_selection(&headers, !conf.no_headers)
+        .map_err(err_msg)?;
+
     let value_pos = args
         .flag_value
-        .single_selection(&headers, !conf.no_headers)?;
+        .single_selection(&headers, !conf.no_headers)
+        .map_err(err_msg)?;
+
     let field_pos_option = args
         .flag_field
         .single_selection(&headers, !conf.no_headers)