+++ src/moonblade/agg/window.rs
@@ -111,6 +111,51 @@ impl RollingWelford {
 }
 
 #[derive(Debug)]
+enum RankingKind {
+    Arbitrary,
+    Dense,
+    CumulativeDistribution,
+    PercentRank,
+    NTile(usize),
+}
+
+impl RankingKind {
+    fn as_str(&self) -> &'static str {
+        match self {
+            Self::Arbitrary => "rank",
+            Self::Dense => "dense_rank",
+            Self::CumulativeDistribution => "cume_dist",
+            Self::PercentRank => "percent_rank",
+            Self::NTile(_) => "ntile",
+        }
+    }
+}
+
+#[derive(Debug)]
+struct Ranking {
+    kind: RankingKind,
+    expr: ConcreteExpr,
+    numbers: Vec<(DynamicNumber, usize)>,
+    output: VecDeque<DynamicNumber>,
+}
+
+impl Ranking {
+    fn new(kind: RankingKind, expr: ConcreteExpr) -> Self {
+        Self {
+            kind,
+            expr,
+            numbers: Vec::new(),
+            output: VecDeque::new(),
+        }
+    }
+
+    fn clear(&mut self) {
+        self.numbers.clear();
+        self.output.clear();
+    }
+}
+
+#[derive(Debug)]
 enum ConcreteWindowAggregation {
     Lead(ConcreteExpr, usize, ConcreteExpr),
     Lag(ConcreteExpr, usize, ConcreteExpr),
@@ -122,7 +167,7 @@ enum ConcreteWindowAggregation {
     RollingSum(ConcreteExpr, RollingSum),
     RollingWelford(ConcreteExpr, WelfordStat, RollingWelford),
     Frac(ConcreteExpr, Sum, Option<usize>),
-    DenseRank(ConcreteExpr, Vec<(DynamicNumber, usize)>, VecDeque<usize>),
+    Ranking(Ranking),
     TotalAggregation(AggregationProgram, DynamicValue),
 }
 
@@ -149,7 +194,7 @@ impl ConcreteWindowAggregation {
     fn requires_total_buffer(&self) -> bool {
         matches!(
             self,
-            Self::Frac(_, _, _) | Self::DenseRank(_, _, _) | Self::TotalAggregation(_, _)
+            Self::Frac(_, _, _) | Self::Ranking(_) | Self::TotalAggregation(_, _)
         )
     }
 
@@ -167,11 +212,16 @@ impl ConcreteWindowAggregation {
                     sum.add(value.try_as_number().map_err(|err| err.specify("frac"))?);
                 }
             }
-            Self::DenseRank(expr, numbers, _) => {
+            Self::Ranking(Ranking {
+                expr,
+                numbers,
+                kind,
+                ..
+            }) => {
                 let value = eval_expression(expr, Some(index), record, headers_index)?;
                 let number = value
                     .try_as_number()
-                    .map_err(|err| err.specify("dense_rank"))?;
+                    .map_err(|err| err.specify(kind.as_str()))?;
 
                 numbers.push((number, numbers.len()));
             }
@@ -186,27 +236,110 @@ impl ConcreteWindowAggregation {
 
     fn finalize_total(&mut self) -> Result<(), SpecifiedEvaluationError> {
         match self {
-            Self::DenseRank(_, numbers, ranks) => {
+            Self::Ranking(Ranking {
+                numbers,
+                output,
+                kind,
+                ..
+            }) => {
                 numbers.sort();
-                ranks.resize(numbers.len(), 0);
+                output.resize(numbers.len(), DynamicNumber::Integer(0));
 
-                let mut rank: usize = 0;
-                let mut last_number: Option<DynamicNumber> = None;
+                let n = numbers.len();
 
-                for (n, i) in numbers.iter() {
-                    match last_number {
-                        None => {
-                            last_number = Some(*n);
+                match kind {
+                    RankingKind::Dense => {
+                        let mut rank: usize = 0;
+                        let mut last_number: Option<DynamicNumber> = None;
+
+                        for (n, i) in numbers.iter() {
+                            match last_number {
+                                None => {
+                                    last_number = Some(*n);
+                                    rank += 1;
+                                }
+                                Some(l) if l != *n => {
+                                    last_number = Some(*n);
+                                    rank += 1;
+                                }
+                                _ => {}
+                            };
+
+                            output[*i] = DynamicNumber::Integer(rank as i64);
+                        }
+                    }
+                    RankingKind::Arbitrary => {
+                        let mut rank: usize = 1;
+
+                        for (_, i) in numbers.iter() {
+                            output[*i] = DynamicNumber::Integer(rank as i64);
                             rank += 1;
                         }
-                        Some(l) if l != *n => {
-                            last_number = Some(*n);
-                            rank += 1;
-                        }
-                        _ => {}
-                    };
+                    }
+                    RankingKind::CumulativeDistribution => {
+                        let mut i: usize = 0;
 
-                    ranks[*i] = rank;
+                        while i < n {
+                            let mut j = i + 1;
+
+                            while j < n && numbers[i].0 == numbers[j].0 {
+                                j += 1;
+                            }
+
+                            let c = j as f64 / n as f64;
+
+                            for k in i..j {
+                                output[numbers[k].1] = DynamicNumber::Float(c);
+                            }
+
+                            i = j;
+                        }
+                    }
+                    RankingKind::PercentRank => {
+                        let mut rank: usize = 1;
+                        let mut i: usize = 0;
+
+                        // Avoiding division by zero
+                        if n > 1 {
+                            while i < n {
+                                let mut j = i + 1;
+
+                                while j < n && numbers[i].0 == numbers[j].0 {
+                                    j += 1;
+                                }
+
+                                let p = (rank - 1) as f64 / (n - 1) as f64;
+
+                                for k in i..j {
+                                    output[numbers[k].1] = DynamicNumber::Float(p);
+                                }
+
+                                rank += j - i;
+                                i = j;
+                            }
+                        }
+                    }
+                    RankingKind::NTile(k) => {
+                        let k = *k;
+
+                        let q = n / k;
+                        let r = n % k;
+
+                        let mut i: usize = 0;
+
+                        'main: for tile in 1..k + 1 {
+                            let size = q + (if tile <= r { 1 } else { 0 });
+
+                            for _ in 0..size {
+                                if i >= n {
+                                    break 'main;
+                                }
+
+                                output[numbers[i].1] = DynamicNumber::Integer(tile as i64);
+                                i += 1;
+                            }
+                        }
+                    }
                 }
             }
             Self::TotalAggregation(program, value) => {
@@ -328,7 +461,9 @@ impl ConcreteWindowAggregation {
                     Some(d) => DynamicValue::from(frac.map(|f| format!("{:.p$}", f, p = d))),
                 })
             }
-            Self::DenseRank(_, _, ranks) => Ok(DynamicValue::from(ranks.pop_front().unwrap())),
+            Self::Ranking(Ranking { output, .. }) => {
+                Ok(DynamicValue::from(output.pop_front().unwrap()))
+            }
             Self::TotalAggregation(_, value) => Ok(value.clone()),
         }
     }
@@ -357,9 +492,8 @@ impl ConcreteWindowAggregation {
             Self::Frac(_, sum, _) => {
                 sum.clear();
             }
-            Self::DenseRank(_, numbers, ranks) => {
-                numbers.clear();
-                ranks.clear();
+            Self::Ranking(ranking) => {
+                ranking.clear();
             }
             Self::TotalAggregation(program, value) => {
                 program.clear();
@@ -374,10 +508,11 @@ fn get_function(name: &str) -> Option<FunctionArguments> {
         "row_number" | "row_index" => FunctionArguments::nullary(),
         "frac" => FunctionArguments::with_range(1..=2),
         "lag" | "lead" => FunctionArguments::with_range(1..=3),
-        "cumsum" | "cummin" | "cummax" | "dense_rank" => FunctionArguments::unary(),
-        "rolling_sum" | "rolling_mean" | "rolling_avg" | "rolling_var" | "rolling_stddev" => {
-            FunctionArguments::binary()
+        "cumsum" | "cummin" | "cummax" | "dense_rank" | "rank" | "cume_dist" | "percent_rank" => {
+            FunctionArguments::unary()
         }
+        "rolling_sum" | "rolling_mean" | "rolling_avg" | "rolling_var" | "rolling_stddev"
+        | "ntile" => FunctionArguments::binary(),
         _ => return None,
     })
 }
@@ -537,12 +672,33 @@ fn concretize_window_aggregations(
                     ConcreteWindowAggregation::Frac(expr, Sum::new(), decimals),
                 ));
             }
-            "dense_rank" => {
+            "dense_rank" | "rank" | "cume_dist" | "percent_rank" => {
                 let expr = concretize_expression(agg.args.pop().unwrap(), headers, None)?;
+
+                let kind = match func_name.as_str() {
+                    "dense_rank" => RankingKind::Dense,
+                    "rank" => RankingKind::Arbitrary,
+                    "cume_dist" => RankingKind::CumulativeDistribution,
+                    "percent_rank" => RankingKind::PercentRank,
+                    _ => unreachable!(),
+                };
 
                 concrete_aggs.push((
                     agg.agg_name,
-                    ConcreteWindowAggregation::DenseRank(expr, Vec::new(), VecDeque::new()),
+                    ConcreteWindowAggregation::Ranking(Ranking::new(kind, expr)),
+                ));
+            }
+            "ntile" => {
+                let expr = concretize_expression(agg.args.pop().unwrap(), headers, None)?;
+                let k = cast_as_usize(&concretize_expression(
+                    agg.args.pop().unwrap(),
+                    headers,
+                    None,
+                )?)?;
+
+                concrete_aggs.push((
+                    agg.agg_name,
+                    ConcreteWindowAggregation::Ranking(Ranking::new(RankingKind::NTile(k), expr)),
                 ));
             }
             _ => unreachable!(),