+++ src/select.rs
@@ -2,7 +2,7 @@ use std::borrow::ToOwned;
 use std::cmp::Ordering;
 use std::convert::TryFrom;
 use std::fmt;
-use std::iter::repeat;
+use std::iter::repeat_n;
 use std::ops;
 use std::str::FromStr;
 
@@ -318,11 +318,11 @@ impl SelectorParser {
     }
 
     fn is_end_of_field(&self) -> bool {
-        self.cur().map_or(true, |c| c == ',' || c == ':')
+        self.cur().is_none_or(|c| c == ',' || c == ':')
     }
 
     fn is_end_of_selector(&self) -> bool {
-        self.cur().map_or(true, |c| c == ',')
+        self.cur().is_none_or(|c| c == ',')
     }
 
     fn bump(&mut self) {
@@ -442,7 +442,7 @@ impl OneSelector {
                 if i < 0 {
                     if i.unsigned_abs() > first_record.len() {
                         Err(format!(
-                            "Selector index {} is out of \
+                            "Column index {} is out of \
                                  bounds. Index must be between -1 \
                                  and -{}.",
                             i,
@@ -455,7 +455,7 @@ impl OneSelector {
                     let i = i as usize;
                     if i >= first_record.len() {
                         Err(format!(
-                            "Selector index {} is out of \
+                            "Column index {} is out of \
                                  bounds. Index must be between 0 \
                                  and {}.",
                             i,
@@ -500,20 +500,20 @@ impl OneSelector {
 
                 if num_found == 0 {
                     Err(format!(
-                        "Selector name '{}' does not exist \
+                        "'{}' does not exist \
                                  as a named header in the given CSV \
                                  data.",
                         s
                     ))
                 } else if sidx < 0 {
                     Err(format!(
-                        "Selector index '{}' for name '{}' is \
+                        "index '{}' for '{}' is \
                                      out of bounds. Must be between -{} and -1.",
                         sidx, s, num_found
                     ))
                 } else {
                     Err(format!(
-                        "Selector index '{}' for name '{}' is \
+                        "index '{}' for name '{}' is \
                                  out of bounds. Must be between 0 and {}.",
                         sidx,
                         s,
@@ -632,7 +632,7 @@ impl Selection {
     }
 
     pub fn indexed_mask(&self, alignment: usize) -> Vec<Option<usize>> {
-        let mut m = repeat(None).take(alignment).collect::<Vec<Option<usize>>>();
+        let mut m = repeat_n(None, alignment).collect::<Vec<Option<usize>>>();
 
         for (j, i) in self.iter().enumerate() {
             if *i < alignment {
@@ -659,7 +659,7 @@ impl Selection {
     }
 
     pub fn contains(&self, i: usize) -> bool {
-        self.0.iter().any(|j| i == *j)
+        self.0.contains(&i)
     }
 
     pub fn subtract(&mut self, other: &Self) {